Make cctor defer Reflection that fails in AOT
This makes it possible that the reflection is never called. In AOT environments where it may still be called (just later) then this too can be avoided in the app by overriding the `MessagePackSecurity.GetHashCollisionResistantEqualityComparer` method and implementing it such that when `T` is `object` it can do something other than invoking the reflection-invoking `ObjectFallbackEqualityComparer` class.

Fixes #1168