Enable publishing of artifacts in multiple passes
When an artifact is not ready yet, but other artifacts are and must be prepared, we can now run `artifacts/_pipelines.ps1` multiple times in the pipeline.
To opt out of an earlier pass, the artifact script need only produce 0 files. After an artifact script returns some files, that artifact will be published and will not be published again.

This is similar to `variables/_pipelines.ps1` behavior, which we already run twice.